redux form and calendar functioning, form needs css
