+++ client/src/components/DatePicker/DatePicker.js
@@ -1,17 +1,34 @@
-import React from "react";
+import React, {Component} from "react";
 import DatePicker from "react-datepicker";
 import moment from "moment";
 
-const RenderDatePicker = props => {
-  console.log(props);
-  return (
-    <div>
-      <DatePicker
-        {...props}
-        onChange={props.onChangeFunc}
-      />
-    </div>
-  );
-};
+
+class RenderDatePicker extends Component {
+  state = {
+    startDate: moment()
+  };
+  changeDate = date => {
+    console.log(date);
+    this.setState({
+      startDate: date
+    });
+  };
+  render () {
+    const {input} = this.props;
+    const formChange = input.onChange;
+    return (
+      <div>
+        <DatePicker
+          dateFormat="DD/MM/YYYY"
+          onChange={(date) => {
+            this.changeDate(date)
+            formChange(date)
+          }}
+          selected={this.state.startDate}
+        />
+      </div>
+    );
+  }
+}
 
 export default RenderDatePicker;
